add print details. set print color, and paper orientation, etc.
git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk@3629 640d5228-2204-0410-8d71-8f46fa6850b4

+++ instruments/usb2d/usb2diag/DataUsb2d64.cc
@@ -93,7 +93,7 @@ void DataUsb2d64::_get2drec() {
      usb2d_rec d;
 
      fread(&h, sizeof(nidas_hdr), 1, _fp);
-     if (h.length==_byte_usb2d_rcd)
+     if (h.length==(unsigned short)_byte_usb2d_rcd)
      {
        fread(&d, h.length, 1, _fp);
        _hdr[c]=h;
